requirements.txt
numpy
matplotlib
pillow
opencv-python-headless
opencv-contrib-python
streamlit
streamlit_option_menu
